chore(STADTPULS-385): rename SWR mutations file
