+++ src/lib/hooks/useUserData/manageSensorsLocally.ts
